requirements.txt
discord.py
aiosqlite
openai
python_dotenv
